Reorganize node-specific code into duffy.nodes
Before, this was scattered over duffy.tasks and duffy.nodes_context.
This is to prepare for pool endpoints which need to access (merged) pool
configuration.
