+++ duffy/nodes/mechanisms/ansible.py
