+++ duffy/nodes/pools.py
