+++ tests/conftest.py
@@ -232,7 +232,7 @@ async def db_async_test_data(db_async_session):
 def test_mechanism():
     """A fixture which provides a no-op "test" mechanism."""
     # This is scoped for the session to ensure that only one mechanism class for type "test" exists.
-    from duffy.tasks.mechanisms import Mechanism
+    from duffy.nodes.mechanisms import Mechanism
 
     class TestMechanism(Mechanism, mech_type="test"):
         pass
